refactor(framework-core): rename modules, functions, & re-org project structure (#89)
