+++ src/github-handler/comment-handler/get-hunk-scope-handler/index.ts
@@ -12,7 +12,4 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-export * from './fork-handler';
-export {branch} from './branch-handler';
-export {commitAndPush} from './commit-and-push-handler';
-export * from './pull-request-handler';
+export {getPullRequestScope} from './remote-patch-ranges-handler';
