+++ test/suggestion-patch.ts
@@ -15,7 +15,7 @@
 import {expect} from 'chai';
 import {describe, it, before, beforeEach} from 'mocha';
 import {setup} from './util';
-import {getValidSuggestionHunks} from '../src/github-handler/comment-handler/suggestion-patch-handler';
+import {getValidSuggestionHunks} from '../src/github-handler/comment-handler/patch-handler/in-scope-hunks-handler';
 import {Range, RawContent} from '../src/types';
 
 before(() => {
